requirements.txt
# Any packages here will be deployed along to the appspot
# Packages in requirements_dev.txt will be installed locally only

unicode-slugify==0.1.1
pytz
voluptuous
git+git://github.com/Workiva/furious.git#egg=furious
google-api-python-client
GoogleAppEngineCloudStorageClient